divide Rule class from system.py
